### 2022-08-31 #4711 @danielskatz
@jhstaudacher - note that JOSS requires reviewed software to have an OSI-approved license (see https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements), which your software doesn't seem to have.
