### 2022-09-01 #4711 @jhstaudacher
Thank you very much for your feedback, @danielskatz. Our software APGG is now under GNU General Public License v3.0.
