### 2022-09-01 #4711 @jhstaudacher
As for potential reviewers for our paper on APGG, I would hereby like to suggest the following six researchers from your list: stulacy, bramvandijk88, jdreo, StefanKohekUMFERI, CFGrote or ieyjzhou.
Thank you very much for your time spent for our submission to JOSS.
