### 2022-10-27 #4711 @Nikoleta-v3
Hey @jhstaudacher 👋🏻 Thank you again for your submission.

Could you please fix the missing DOI for the following references https://github.com/openjournals/joss-reviews/issues/4711#issuecomment-1293686938?
