### 2022-10-27 #4711 @Nikoleta-v3
👋🏻 @ieyjzhou @dglmoore @junliu2050 would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "APGG - A Modular C++ Framework for Asymmetric Public Goods Games". You can find more information at the top of this Github issue 🆙
