### 2022-10-27 #4711 @junliu2050
Sorry, not in my expertise.

Sent from my iPhone

> On Oct 27, 2022, at 10:17 AM, Nikoleta Glynatsi ***@***.***> wrote:
>
> ﻿
> 👋🏻 @ieyjzhou @dglmoore @junliu2050 would any of you be willing to review this submission for JOSS?
>
> We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
> The submission I'd like you to review is titled: "APGG - A Modular C++ Framework for Asymmetric Public Goods Games". You can find more information at the top of this Github issue 🆙
>
> Please let me know if you're available 😄 Thank you!
>
