+++ paper/paper.bib
@@ -6,6 +6,8 @@
   number={3},
   pages={193--212},
   year={1995},
+  doi{10.1162/evco.1996.4.2.113},
+  url = {https://doi.org/10.1162/evco.1996.4.2.113},
   publisher={[Champaign, IL, USA: Complex Systems Publications, Inc., c1987-}
 }
 @article{hardin1968tragedy,
@@ -16,6 +18,8 @@
   number={3859},
   pages={1243--1248},
   year={1968},
+  doi{10.1126/science.162.3859.1243},
+  url = {https://doi.org/10.1126/science.162.3859.1243},
   publisher={American Association for the Advancement of Science}
 }
 @article{mcginty2013public,
@@ -26,6 +30,8 @@
   number={4},
   pages={1159--1177},
   year={2013},
+  doi{10.1007/s00355-012-0658-2},
+  url = {https://doi.org/10.1007/s00355-012-0658-2},
   publisher={Springer}
 }
 @article{hintze2020inclusive,
@@ -36,6 +42,8 @@
   number={1},
   pages={1--8},
   year={2020},
+  doi{10.1038/s41598-020-79731-y},
+  url = {https://doi.org/10.1038/s41598-020-79731-y},
   publisher={Nature Publishing Group}
 }
 @article{adami2016evolutionary,
@@ -45,6 +53,8 @@
   volume={19},
   pages={1--26},
   year={2016},
+  doi{10.1016/j.plrev.2016.08.015},
+  url = {https://doi.org/10.1016/j.plrev.2016.08.015},
   publisher={Elsevier}
 }
 @inproceedings{bohm2017mabe,
@@ -53,6 +63,8 @@
   booktitle={ECAL 2017, the Fourteenth European Conference on Artificial Life},
   pages={76--83},
   year={2017},
+  doi{10.1162/isal_a_016},
+  url = {https://doi.org/10.1162/isal_a_016},
   organization={MIT Press}
 }
 @article{fehr2002altruistic,
@@ -63,6 +75,8 @@
   number={6868},
   pages={137--140},
   year={2002},
+  doi{10.1038/415137a},
+  url = {https://doi.org/10.1038/415137a},
   publisher={Nature Publishing Group}
 }
 @article{hintze2015punishment,
@@ -73,6 +87,8 @@
   number={4},
   pages={046005},
   year={2015},
+  doi{10.1088/1478-3975/12/4/046005},
+  url = {https://doi.org/10.1088/1478-3975/12/4/046005},
   publisher={IOP Publishing}
 }
 @article{smith2007rank,
@@ -83,6 +99,8 @@
   number={5},
   pages={753--765},
   year={2007},
+  doi{10.1007/s00265-006-0305-y},
+  url = {https://doi.org/10.1007/s00265-006-0305-y},
   publisher={Springer}
 }
 @article{lenski2017experimental,
@@ -93,6 +111,8 @@
   number={10},
   pages={2181--2194},
   year={2017},
+  doi{10.1038/ismej.2017.69},
+  url = {https://doi.org/10.1038/ismej.2017.69},
   publisher={Nature Publishing Group}
 }
 
@@ -100,6 +120,8 @@
   title={Evolutionary Computational Modeling in a 3D Physics Environment},
   author={Richter, David Jona},
   year={2019},
+  doi{10.13140/RG.2.2.27100.72322/2},
+  url = {https://doi.org/10.13140/RG.2.2.27100.72322/2},
   type={Bachelor's Thesis},
   school={Kempten University}
 }
