### 2022-10-28 6feb5ca
Fixed syntax issue with DOIs
