+++ paper/paper.bib
@@ -63,8 +63,8 @@
   booktitle={ECAL 2017, the Fourteenth European Conference on Artificial Life},
   pages={76--83},
   year={2017},
-  doi = {10.1162/isal_a_016},
-  url = {https://doi.org/10.1162/isal_a_016},
+  doi = {10.7551/ecal_a_016},
+  url = {https://doi.org/10.7551/ecal_a_016},
   organization={MIT Press}
 }
 @article{fehr2002altruistic,
