### 2022-10-28 #4711 @Nikoleta-v3
👋🏻 @raybecker @sylvaticus would either of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "APGG - A Modular C++ Framework for Asymmetric Public Goods Games". You can find more information at the top of this Github issue 🆙
