### 2022-10-29 #4711 @Nikoleta-v3
> I'm in... which is the expected timing ?

We aim for reviews to be completed within about 2-4 weeks.

> Thanks @sylvaticus. @Nikoleta-v3, you don't need both of us, correct?

That is correct 😃  we already have one reviewer.

> Just to be clear, I'm an economist but experimental and behavioural economics are not my fields, so feel free to drop me out of the list...

APGG is a tool for numerical simulations of the public goods game. I believe the submission will benefit from a reviewer that has a good knowledge of C++ and software development more than someone familiar with its game theoretic aspects.  Let me know if that sounds good to you :)

And thank you both for your replies 🙏🏻
