### 2022-10-31 #4711 @sylvaticus
ok, I leave the task to @raybecker then.. my C++ is a bit rusty..  it's a while I don't use it, as I moved almost everything to Julia nowadays...

Just a few comments/suggestions to the author package @jhstaudacher  if I can:
- add a couple of sentences after the logo on what your program does;
- create a proper documentation (there is no link to the "our wiki" page referred in the README);
- linked to the previous point, the code I saw has no comments. In C++ you have tools like [Doxygen](https://doxygen.nl/) to help you structure the documentation that you write close to the code as "comment" and automatically extract the API. Use them :-) ;
- you shouldn't need to register to download an open source software, check the "release" method;
- on a quick look on the repository I haven't see any [unit test](https://en.wikipedia.org/wiki/Unit_testing). They are really useful, and [some people even write first the test and then the implementation](https://en.wikipedia.org/wiki/Test-driven_development) !
