### 2022-11-07 #4711 @Nikoleta-v3
👋🏻 @manmeetkaurbaxi & @gabett would either of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "APGG - A Modular C++ Framework for Asymmetric Public Goods Games". You can find more information at the top of this Github issue 🆙
