### 2022-11-14 #4711 @Nikoleta-v3
Hey @gabett 👋🏻  thank you very much for your reply!

I realise that this is not your field of expertise. I believe the submission will benefit from a reviewer that has a good knowledge of C++ and software development more than someone familiar with its game theoretic aspects.

Given this, do you think that it could be something you can review? 😄

"No, I rather prefer not to" is a perfectly fine answer!
