### 2022-11-17 #4711 @Nikoleta-v3
Hey @MikeLydeamore & @mstimberg  would either of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "APGG - A Modular C++ Framework for Asymmetric Public Goods Games".
I realise that this submission is not in your field of expertise. However, I believe the submission will benefit from a reviewer that has a good knowledge of C++ and software development more than someone familiar with its game theoretic aspects. You can find more information about the submission at the top of this Github issue 🆙
