### 2022-11-18 #4944 @editorialbot
[REVIEW]: APGG - A Modular C++ Framework for Asymmetric Public Goods Games
**Submitting author:** @jhstaudacher (<a href="http://orcid.org/0000-0002-0619-4606">Jochen Staudacher</a>)
**Repository:** https://github.com/APGG-Lab/APGG
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.2
