### 2022-11-18 #4944 @Nikoleta-v3
@jhstaudacher, you might recall that [sylvaticus](https://github.com/sylvaticus) , even though they could not review the submission, left some useful suggestion/comments on https://github.com/openjournals/joss-reviews/issues/4711#issuecomment-1319729427. I am not sure if you have addressed these. Just in case I am pasting them here:

Just a few comments/suggestions to the author package jhstaudacher  if I can:
- add a couple of sentences after the logo on what your program does;
- create a proper documentation (there is no link to the "our wiki" page referred in the README);
- linked to the previous point, the code I saw has no comments. In C++ you have tools like [Doxygen](https://doxygen.nl/) to help you structure the documentation that you write close to the code as "comment" and automatically extract the API (perhaps even other ones better nowadays). Use them :-) ;
- you shouldn't need to register to download an open source software, check the "release" method;
- on a quick look on the repository I haven't see any [unit test](https://en.wikipedia.org/wiki/Unit_testing). They are really useful, and [some people even write first the test and then the implementation](https://en.wikipedia.org/wiki/Test-driven_development) !
