### 2022-11-20 #4944 @jhstaudacher
@nikoleta-v3 Thank you very much for reminding us of the precious suggestions and comments on our software APGG that sylvaticus provided three weeks ago. We will address these issues in the next release of our software APGG once the reviews are completed. Right now, I feel it is preferable not to change the code while the reviewers are assessing the software. Thus we can avoid confusion and misunderstanding. We will create a new release of APGG -- and revise our paper for JOSS -- addressing all suggestions from the reviewers (including the comments from sylvaticus) after we have the reviews. Thank you very much again for all your efforts.
