### 2022-12-09 APGG-Lab/APGG#2 @mstimberg
Restructure documentation
[Issue opened as part of review openjournals/joss-reviews#4944]

I would suggest restructuring the documentation (i.e, the [wiki](https://github.com/APGG-Lab/APGG/wiki)) a bit to make it more accessible for users:
* The order in which the content is presented should follow the workflow, i.e. first installation (from release or source), then setting up the configuration, then running, etc. Or at least it should start with a "Getting Started" section
* You might reconsider what information goes into the README file – currently, it is partly redundant with the wiki, but not sufficient to actually get started (it does not mention how to run simulations and how results are stored). Also, it only describes how to install from source on Windows, whereas the wiki also has information for Linux.
* The section title "How to setup Batch runs" suggests this is an alternative way of running experiments, where to me it seems that editing the `configs.csv` file is the only way to configure experiments (i.e., it is also needed for a single run).
* I'd more clearly separate information such as "class structure", which is mostly relevant to users that want to extend the C++ code, from information relevant for "non-expert users".
* Not sure an empty FAQ is particularly helpful ;)
