### 2022-12-09 APGG-Lab/APGG#3 @mstimberg
Revisit release distribution
[Issue opened as part of review openjournals/joss-reviews#4944]

As already mentioned by another reviewer, the release link in the README file links to a non-public website. The tagged releases on the GitHub repository only provide the source archive. Consider either automating the release builds (e.g. via GitHub Actions), or manually upload the archive to a GitHub release.
