### 2022-12-09 APGG-Lab/APGG#4 @mstimberg
Provide clear instructions how to run/analyze an example simulation
[Issue opened as part of review openjournals/joss-reviews#4944]

The repository contains an example config file in [`build/configs.csv`](https://github.com/APGG-Lab/APGG/blob/master/build/configs.csv) which I think is a coarse-grained version of the simulation included in the paper. The documentation should use this (or another) example to give step-by-step instructions how to run a simulation and analyze the results. Currently, the wiki only states "Analyse the experiment by dropping the experiment folder on one of the python scripts in the experiments folder" – I would strongly recommend augmenting this by a description of what each script does. The Python scripts also depend on external libraries (matplotlib, pandas), these dependencies should be stated.
