### 2022-12-09 APGG-Lab/APGG#5 @mstimberg
Improve documentation for config values
[Issue opened as part of review openjournals/joss-reviews#4944]

Both the README and the wiki contain a documentation of the configuration values, but I feel that the information is not displayed in the most convenient way. In particular, the wiki states that the "config can be splitted in 3 big chunks", but a category like "ClassTypes" could use additional explanation. Also, it might be clearer to have the categories as subheadings, instead of stating a "Category" for each entry.
