### 2022-12-09 APGG-Lab/APGG#6 @mstimberg
Should "Visualizer" appear in the class structure?
[Issue opened as part of review openjournals/joss-reviews#4944]

The class structure diagram in the wiki and the paper contains the `Visualizer` class, but if I understand correctly, there is no such class in the C++ code, instead visualization is typically done with external tools (e.g. Python scripts). It seems odd to include it in the UML diagram in that case.
