### 2022-12-09 #4944 @mstimberg
As a reminder: the topic is outside my field of expertise, so I cannot really judge, say, the appropriateness of the methods or the relevance to the field. I will therefore (mostly) constrain my comments on the technical aspects.

I was able to successfully compile and run the software (on Linux). IMO, the C++ code is of good quality and rather well-structured and readable. I agree with the comments by sylvaticus that it would be good to use some API documentation tool and/or unit testing, but from my point of view this is not strictly necessary for the paper/software acceptance into JOSS. However, I feel that there needs to be better step-by-step documentation for how to verify the correct working of the software, e.g. by reproducing Figure 4 of the manuscript. I've opened this and a few other issues directly over in the APGG repository (APGG-Lab/APGG#2, APGG-Lab/APGG#3, APGG-Lab/APGG#4, APGG-Lab/APGG#5, APGG-Lab/APGG#6).

I also have a few comments on the manuscript, I'll directly include them in this comment here:
* As I mentioned earlier, I am not from the field, but I *do* know the (in)famous "Tragedy of the commons" essay by Hardin.  I am sure the authors are aware that the essay is highly problematic, to say the least. As e.g. [Matto Mildenberger noted](https://blogs.scientificamerican.com/voices/the-tragedy-of-the-tragedy-of-the-commons/), "[Hardin] was a racist, eugenicist, nativist and Islamaphobe—plus his argument was wrong". I don't have a clear idea of what the best approach here would be, but I strongly feel that citing this paper/author without comment is not appropriate nowadays.
* Figure 4 is claimed to replicate Figure 6 from Hintze & Adami (2015), but I have trouble mapping the results between the two figures. Also, shouldn't the "percentile of agents" (rather "percentage"?)  for "cooperators" and "punishers" add up to 100? The authors state that "For each datapoint along the axis of the synergy factor 10 replicate experiments
were run", but do I understand correctly that the figure only shows the results from a single simulation? Wouldn't it make sense to use e.g. error bars or intervals to use all the information available? Finally, either the manuscript or the documentation should give clear instructions how to reproduce the figure.

Minor issues/comments:
* In l.29/30, the references seem to be for the MABE framework. Is this just meant as an example of a modular framework? Or APGG is related to MABE in some way?
* the citations in `paper.md` should use e.g. `[@mcginty2013public,@hintze2020inclusive]` instead of `[@mcginty2013public],[@hintze2020inclusive]` to correctly display multiple citations. Also, try reformatting the reference in l. 130.
* Figure 4: I wouldn't claim that a single successful replication *proves* that "APGG does work as intended"
* l. 103: I think it should be made clearer that `itIsNotReallyAGroupLevelPayOffCalculator` (what a name!), is an example for an extension, and not part of APGG (in case I understood that correctly, of course).
