### 2022-12-18 #4944 @jhstaudacher
Hey @Nikoleta-v3, thank you very much for your message from last Friday, December 16.
Yes, I forwarded the very helpful review by Marcel Stimberg (mstimberg) from Friday, December 16, 17:34 (CET), as well as the very helpful comments by Yanjie Zhou (ieyjzhou) from Friday, December 16, 18:09 (CET), to my coauthors and we already discussed the feedback we received. We will be very happy to revise our software APGG and our paper submission accordingly. Still, we are not certain whether the first round of reviews has actually been completed or whether there will be some kind of formal "decision letter" from you as our editor.
I am sorry if this question appears obtuse, but for my coauthors and me it is our first submission to JOSS. Thank you very much in advance for your reply and your efforts in editing our submission to JOSS.
