### 2022-12-19 #4944 @Nikoleta-v3
Hey @jhstaudacher  👋🏻

> I am sorry if this question appears obtuse, but for my coauthors and me it is our first submission to JOSS.

Not a problem at all! 😊

In general you won’t receive any formal decision letter from me or from JOSS. Everything (reviews, acceptance, rejection) will happen on this issue.

The format that JOSS has for the reviews is that reviewers write their comments here, or open issues on the project’s repository, and the authors address them. I won’t intervene, except if I believe an intervention is needed. The reason for this format is to keep the discussion between you and the reviewers ongoing. This way you can have a real conversation with them.

Regarding [mstimberg](https://github.com/mstimberg)'s review: From their checklist (https://github.com/openjournals/joss-reviews/issues/4944#issuecomment-1344428173) there are only a few points that have not been ticked. They have raised their issues with these points by opening issues on.
I would now start addressing them. Ideally for each issue you can have a different commit or even a pull request if the issue requires a lot of work.  If you have any questions you can reply on the issue for clarification. Once you are done addressing the issue, comment on the issue that you are done and ideally also include a link to the commit/pull request.

Regarding [ieyjzhou](https://github.com/ieyjzhou)'s review: I can see that they have also raised some issues. I would address them and comment here how you resolved them (again with links to commits/pull requests). @ieyjzhou has to generate their checklist.

Once both reviewers have completed their checklist we can move forward. Please let me know if everything makes sense or if you have any further questions!
