### 2023-01-07 270ebdc
Reverted back to old citation syntax after finding that the new one omits the paper listed first.
