### 2023-01-07 APGG-Lab/APGG#5 @JDatPNW
https://github.com/APGG-Lab/APGG/wiki/ConfigValues/64a785a08081e0aba6f44766e2ac7b281a0ee478
