### 2023-01-07 0b3ba56
Added Getting started to README.md.
