### 2023-01-13 #4944 @jhstaudacher
Hey @Nikoleta-v3, thank you very much for your message from January 13 and your New Year greetings from January 3. Please allow me to start my reply reciprocating your good wishes for 2023.
In terms of the progress on our revision, we already addressed four issues (APGG#3 to APPG#6). Our coauthor Falk Hübner (Venku628) replied to these four issues last Saturday, January 7, and closed them.
There is still some work to be done before our resubmission -- in particular, restructuring the documentation for APGG and addressing the points brought up by sylvaticus and ieyjzhou.
Of course, we will definitely not forget to address all the "Minor issues/comments" from the review by Marcel Stimberg (mstimberg) and we are also generating an improved version of Figure 4 in our paper.
In short: My coauthors and me are half way through and we still need a bit more time to address all the points raised by the reviewers.
Thank you very much again for all your efforts in editing our paper.
