### 2023-02-20 #4944 @jhstaudacher
Hey @Nikoleta-v3, thank you very much for your message from Friday, February 17. My coauthors and me are still working on restructuring the documentation (APGG#2) and some seemingly smaller issues. Hopefully, we shall be able to resubmit next month.
Thank you very much for your patience and thank you very much again for all your efforts in editing our paper.
