### 2023-02-20 88d959f
Updated Paper to use correct Citation Syntax as well as proper Figure References
