### 2023-02-20 168e898
Added additional information and helpful references to the README file.
