+++ README.md
@@ -2,7 +2,7 @@
 
 ![Logo](https://mirko.dev/apgg_logo_full.png)
 
-Code for asymetric public goods games
+APGG is a Framework that allows users to run, simulate and analyze Public Goods Games Experiments. APPG is written in a way that should allow users of all levels of experience to use, utalize and modify the Framework as desired.
 
 ## Installation
 * Download the release from the [Release section ](https://github.com/APGG-Lab/APGG/tags)
@@ -20,7 +20,7 @@ For more information, please refer to our [Wiki page.](https://github.com/APGG-L
 ## Getting started
 1. Adjust the values in the main function of buildConfig.py to whatever values are suitable for your experiment (see Config Setup or [Wiki entry](https://github.com/APGG-Lab/APGG/wiki/ConfigValues)). By adding addIteration calls you can specify entire ranges of values for a config value that the script writes for you automatically. With multiple calls to this function every possible permutation of all values ranges is written. All these experiment setups will be run in sequence.
 2. Run the buildConfig.py to generate the config.csv.
-3. Run the executable. 
+3. Run the executable.
 4. After the conclusion of the experiments in order to visualize the data first run combineExperiments.py on the generated folder.
 5. Run plot.py to get a chart of your data. Adjust as needed.
 
@@ -47,11 +47,11 @@ For more information on this topic, please refer to our Wiki pages on [Config Va
 |mutatorType | int  (0-1) | 0 = RandomMutator / 1 = ThresholdMutator (recommended by Dr. Hintze) |
 |mutationRate | int  (0-100) | Mutationrate. 0 = no Mutation, 100 = every Organism will mutate every round |
 |showAllGenerations | int  (0-1)  | Log all generations to std. out |
-|archiveData | int  (0-1) | Archive data to csv | 
-|visualize | int (0-1) | Open python3 script after experiment (unused) | 
+|archiveData | int  (0-1) | Archive data to csv |
+|visualize | int (0-1) | Open python3 script after experiment (unused) |
 |folderName| string  | name of the experiment
 |logSuffix | string | Suffix for the experiment .csv file |
-|timeToFile | int  (0-1) | Add timestamp to file | 
+|timeToFile | int  (0-1) | Add timestamp to file |
 |timeToFolder | int (0-1) | Add timestamp to folder |
 |consoleOutExponent | int  | Multiplicator for interval output |
 
