### 2023-03-19 28987c9
Enlarged flow chart image in the paper file in hopes that it will fix existing formatting issues
