### 2023-03-19 817d1c7
Decreased flow chart image size in the paper file in hopes that this time it will actually fix existing formatting issues
