### 2023-03-19 5ade886
Added comments to all classes to clarify their purpose and their basic functionality. Functions did not receive comments due to them being self explanatory by their name.
