+++ APGG/repopulators/ProportionateRepopulator.h
@@ -4,6 +4,10 @@
 
 namespace APGG {
 
+    /**
+    * Basic implementation of the Repopulator base class that repopulates eliminated Organisms with parent candidates chosen randomly
+    * and then decided upon by the payoff they might provide.
+    */
     class ProportionateRepupoluator : public Repopulator
     {
     public:
