+++ APGG/selectors/EliteSelector.h
@@ -3,7 +3,10 @@
 
 namespace APGG {
 
-
+    /**
+    * Basic implementation of the Selector base class that selects Organisms from a Grid based on their order after
+    * sorting the Grid by fitness. 
+    */
     class EliteSelector : public Selector
     {
     public:
