### 2023-03-24 315e0a1
Update paper.md
edit figure 6
