### 2023-03-24 b412e63
Update paper.md
figure caption update
