+++ paper/paper.md
@@ -112,7 +112,7 @@ The computer model runs using the following steps:
 To illustrate that the software is capable of reproducing scientific results, we repeated an experiment from Hintze and Adami (2015) [@hintze2015punishment]. In that experiment the synergy factor in the public goods game was varied and the outcome of evolution, given that factor, determined. When the synergy factor is low, we expect defectors to win, while cooperators should thrive when synergy is high. However, in this variant of the game, agents can also punish, and punishment can modulate the response of agents to the synergy factor. The expectation was, that punishment should allow agents to cooperate at lower synergy factors, and one should observe punishment to increase around this critical point. Interestingly, agents probabilities to cooperate and to punish evolved as expected, and punishment lowered the critica point, but the chance to defect went from 0.0 to 0.5 (drifting). Here we replicated that exact result (see \autoref{fig:Figure6}).
 
 ![This Figure displays the same result as [@hintze2015punishment], therefore proving that APGG does work as intended.\label{fig:Figure6}](img/Figure6.png){ width=100% }
-[Test]
+Figure Caption: Evovled agent behavior (y-axes) in the public goods game with punishment for different synergy factors (x-axis). The evolved likelihood to cooperate (in black, left y-axis) and to punish (in red, right y-axis) is shown for 20 replicate experiments per tested synergy factor value. Shadows in red and black show the 95% confident intervalls respectively. The expected critical point for agents not being able to punish is at a synergy factor of 5.0, and here we find this critical point to be shifted to a lower level as expected.
 
 
 ## Conclusion
