### 2023-03-24 c1a6bba
Improving figure caption for figure 4
