### 2023-03-24 7ce7274
Reference the work by Hardin (1968) in a different way, including an additional reference to the recent work by Mildenberger (2019) as kindly suggested by reviewer mstimberg
