+++ paper/paper.bib
@@ -22,6 +22,13 @@
   url = {https://doi.org/10.1126/science.162.3859.1243},
   publisher={American Association for the Advancement of Science}
 }
+@misc{mildenberger2019tragedy,
+  title = {The Tragedy of the Tragedy of the Commons},
+  author={Mildenberger, Matto},
+  howpublished = {\url{https://blogs.scientificamerican.com/voices/the-tragedy-of-the-tragedy-of-the-commons/}},
+  year ={2019},
+  note = {Accessed: 2023-03-24}
+}
 @article{mcginty2013public,
   title={Public goods provision by asymmetric agents: experimental evidence},
   author={McGinty, Matthew and Milam, Garrett},
