### 2023-03-24 #4944 @jhstaudacher
Hey @Nikoleta-v3, thank you very much for your very kind responses to my three questions from Thursday night.
In particular, thank you very much for encouraging me to use the editorialbot for updating our paper ... I used http://whedon.theoj.org/ previously in order to avoid generating email messages and disturb you.
The version I just generated is the one I am now sending -- together with the final draft of the point-by-point responses to the reviewers -- to my coauthors for approval (or last changes). I shall resubmit as soon as I have "green light" by my coauthors.
Thank you very much for your patience and all your kindness. Have a wonderful weekend.
