### 2023-04-06 1c9ac28
Merge two references in line 43 of our paper as suggested by reviewer Marcel Stimberg.
