### 2023-04-06 e173928
Merging the two paragraphs for the asymmetric payoff calculator and the GroupLevelPayoffCalculator in our paper for JOSS as suggested by reviewer Marcel Stimberg.
