### 2023-04-17 7a2257b
Add funding acknowledgement for Jochen Staudacher for our paper for JOSS.
