### 2023-04-17 7d63a2a
Update text around figure 4 in JOSS paper according to suggestion by reviewer Marcel Stimberg. Figure 4 is not an exact replication of figure 6 in Hintze & Adami (2015), just a qualitative one.
