### 2023-04-20 9a501d9
JOSS-Paper: Address point by reviewer Marcel Stimberg and point out that our Figure 4 reproduces the figure from Hintze & Adami (2015) qualitatively using a smaller number of replicates (thus allowing for a more meaningful visualization of confidence intervals).
