### 2023-04-20 #4944 @jhstaudacher
@Nikoleta-v3 thank you very much again for editing our paper. With approval from my coauthors, I am hereby resubmitting.

RESPONSE TO THE REVIEWER:

accompanying the resubmission of the paper

'APGG - A Modular C++ Framework for Asymmetric Public Goods Games'

by Mirko Rosenthal, David J. Richter, Falk Hübner, Jochen Staudacher and Arend Hintze

@mstimberg Thank you very much again for your second review of our paper from April 6 and your very helpful suggestions and comments.

>
> In particular, it is still unclear to me how to
> reproduce Figure 4:
>
> The "Getting started" section in the README/wiki states to
> 1) "Adjust the values in the main function of buildConfig.py
> to whatever values are suitable for your experiment […]" and
> 2) "Run the buildConfig.py to generate the config.csv."
> This seems to be very general ("your experiment") and not
> specific to the example shown in Figure 4. It looks as if
> the buildConfig.py file contains the settings to reproduce
> the values for Figure 4 by default, but this should be stated
> explicitly. If I run the unchanged buildConfig.py, it creates
> a config.csv file, but trying to run this file with the apgg
> binary does not work for me, it loads the experiments but then
> emits the warning [APGG Warning] You requested a non existing
> config key "timeToFolder". Using default: "" for each of them
> and does not generate any results. If I run the apgg binary
> with the config.csv value from the repository (i.e. without
> running buildConfig.py first), it will run the simulations
> successfully, but I cannot run combineExperiments.py with the
> results, since apparently this script also expects a setup.npy
> file. This file seems to be generated by buildConfig.py, but is
> not mentioned anywhere. As I stated before in APGG-Lab/APGG#4,
> it would be important to include at least some details about what
> these Python scripts do, and which dependencies they need.
>

Thank you very much for your precious suggestions and observations. Our wiki now contains both a separate page "Plotting Data" (see https://github.com/APGG-Lab/APGG/wiki/Plotting-Data ) and a separate page "Proof of Concept Replication" (see https://github.com/APGG-Lab/APGG/wiki/Proof-of-Concept-Replication ) with detailed instructions on how to reproduce Figure 4 from our paper.

We also updated the following three pages in our wiki reflecting your suggestions for improvements

https://github.com/APGG-Lab/APGG/wiki/CSV-Batch-Experiments
https://github.com/APGG-Lab/APGG/wiki/Run-APGG
https://github.com/APGG-Lab/APGG/wiki/Getting-Started

and fixed a problem in our code in the file "Archiver.cpp".

>
> Could you also comment a bit on how the Figure 4
> in this paper "replicated that exact result" from
> Hintze & Adami (2015)? Again, I'm not an expert in
> this domain, but I thought the Figure 4 from the
> JOSS manuscript was meant to replicate Figure 6
> from Hintze & Adami (2015), but this does not seem
> to be the case (P_C and P_P raise steeply at
> synergy 4 in Hintze & Adami (2015) while they raise
> less steeply starting at synergy ~3 in the JOSS manuscript).
>

Thank you very much for being so observant. We are very grateful since we should not have written that Figure 4 in our paper was an exact replication of the result from Hintze & Adami (2015). Instead, Figure 4 reproduces the same result qualitatively using a smaller number of replicates (-- the latter also allows for a meaningful visualization of confidence intervals which you suggested in your first review).

>
> Regarding the releases (APGG-Lab/APGG#3), the link in the
> README now points to GitHub's release section, but these
> releases are both very old (latest release from Oct 6, 2020),
> and do not contain any pre-compiled binaries. According to
> the wiki, building from source is optional, but I don't see
> any other option at the moment.
>

We created a new release (1.1.1) of APGG including binaries on April 12.

Finally, let us address the three minor issues:

>
> The wiki on config values has broken/missing links
> for World (should be World-Class ?) and MatchupGenerator
>

Thank you very much. We fixed the links.

>
> The two citations in l. 43 should be merged.
>

Done.

>
> L.99–107: this might be just an issue of formatting,
> but starting with "There are payoff calculators for
> different scenarios:" it only introduces a single payoff
> calculator (asymmetric payoff). The GroupLevelPayoffCalculator
> (should it be AsymmetricPayoffCalculator as well?) is
> then described in a new paragraph.
>

Thank you very much for your observation. We adjusted the formatting.
