### 2023-04-20 6f93c44
JOSS-Paper: I had overlooked a minor point in reviewer Marcel Stimberg's suggestions. For consistency, we should speak about "AsymmetricPayoffCalculator" rather than "Asymmetric Payoff calculator".
