### 2023-04-25 33e7ee8
Adjust minor issue in paper as suggested by reviewer M. Stimberg ...
