### 2023-06-12 b2e7828
Update buildConfig.py
