### 2023-06-12 98b68ba
Fixed buildConfig.py missing trailing semi colon for linux
