+++ paper/paper.md
@@ -101,15 +101,25 @@ The computer model runs using the following steps:
 
 ## Proof of Concept
 
-| Parameters          | $r$              |$\mu$        | Pop. Size       |Group Size |  $\gamma$       |  $\beta$         |
-|---------------------|----------------|---------------|-----------------|-----------|-----------------|------------------|
-| Value:              |   1.5 to 7.0   |  0.02         |     1000        |    5      | 0.2             |   0.8            |
+| Parameters          | $r$              |$\mu$        | Pop. Size       |Group Size |  $\gamma$       |  $\beta$         |  $i$           |
+|---------------------|----------------|---------------|-----------------|-----------|-----------------|------------------| ---------------|
+| Value:              |   1.5 to 7.0   |  0.02         |     1000        |    5      | 0.2             |   0.8            |   1            |
 
-**Table 1**: Parameters used to replicate Proof of Concept Experiments, where $r$ is the Synergy Factor, $\mu$ is the Mutation Rate, $\gamma$ is the Punishment Cost and $\beta$ is the Punishment Fine. The experiment was run for 500,000 generations with 10 replicates. The Random Mutator was picked, as was the Random Selector. The Repopulator used was the Proportionate Repopulator.
+**Table 1**: Parameters used to replicate our first Proof of Concept Experiment, where $r$ is the Synergy Factor, $\mu$ is the Mutation Rate, $\gamma$ is the Punishment Cost, $\beta$ is the Punishment Fine and $i$ is the payoff individualism (with $i = 1$ denoting the maximal level level of selfishness and thus the classical public goods game). The experiment was run for 100,000 generations with 100 replicates. The Random Mutator was picked, as was the Random Selector. The Repopulator used was the Proportionate Repopulator.
 
-To illustrate that the software is capable of reproducing scientific results, we repeated an experiment from [@hintze2015punishment]. In that experiment the synergy factor in the public goods game was varied and the outcome of evolution, given that factor, determined. When the synergy factor is low, we expect defectors to win, while cooperators should thrive when synergy is high. However, in this variant of the game, agents can also punish, and punishment can modulate the response of agents to the synergy factor. The expectation was, that punishment should allow agents to cooperate at lower synergy factors, and one should observe punishment to increase around this critical point. Interestingly, agents' probabilities to cooperate and to punish evolved as expected, and punishment lowered the critical point, but the chance to defect went from 0.0 to 0.5 (drifting). 10 replicates (rather than 100 in the original experiment from [@hintze2015punishment]) are sufficient to reproduce that result qualitatively. \autoref{fig:Figure6} confirms both the phenomenon of drifting and the shift of the critical point from 5.0 (without punishment) down to a lower level around 4.0 (with punishment).
+To illustrate that the software is capable of reproducing scientific results, we first repeated an experiment from [@hintze2015punishment]. In that experiment the synergy factor in the public goods game was varied and the outcome of evolution, given that factor, determined. When the synergy factor is low, we expect defectors to win, while cooperators should thrive when synergy is high. However, in this variant of the game, agents can also punish, and punishment can modulate the response of agents to the synergy factor. The expectation was, that punishment should allow agents to cooperate at lower synergy factors, and one should observe punishment to increase around this critical point. Interestingly, agents' probabilities to cooperate and to punish evolved as expected, and punishment lowered the critical point, but the chance to punish went from 0.0 to 0.5 (drifting). \autoref{fig:Figure6} confirms both the phenomenon of drifting and the shift of the critical point from 5.0 (without punishment) down to a lower level around 4.0 (with punishment).
 
[…]
