### 2023-06-19 9dbe662
Write more about critical points to explain better that our graphs are correct and meaningful.
