+++ paper/paper.md
@@ -113,7 +113,7 @@ Let us look at the shift of the critical point from 5.0 (without punishment) dow
 \begin{align} \label{criticalPointWithP}
 r_{C} = i (1 - \beta \rho_{P}) (k+1) - i + 1  - (1- i) (\beta + \gamma) \rho_{P}.
 \end{align}
-For example, for $\rho_{P}=0.4$ and our setting of parameters, expression (\ref{criticalPointWithP}) would for predict $r_{C} = 1 \cdot (1 - 0.8 \cdot 0.4) \cdot 5 = 3.4 without mutations. As expected we observe a transition from defection to cooperation for synergy factors between 2.5 and 4 in \autoref{fig:Figure6}.
+For example, for $\rho_{P}=0.4$ and our setting of parameters, expression (\ref{criticalPointWithP}) would for predict $r_{C} = 1 \cdot (1 - 0.8 \cdot 0.4) \cdot 5 = 3.4$ without mutations. As expected we observe a transition from defection to cooperation for synergy factors between 2.5 and 4 in \autoref{fig:Figure6}.
 
 ![Evolved agent behavior (y-axes) in the public goods game with punishment for different synergy factors (x-axis). The evolved likelihood to cooperate (in black, left y-axis) and to punish (in red, right y-axis) is shown for 100 replicate experiments per tested synergy factor value. Shadows in red and black show the 95% confidence intervals respectively. The expected critical point for agents not being able to punish is at a synergy factor of 5.0, and here we find this critical point to be shifted to a lower level as expected. \label{fig:Figure6}](img/Figure6.png){ width=100% }
 
