### 2023-06-20 9425c44
Further improvements for resubmission of JOSS paper.
