### 2023-06-20 4d2ba35
Further improvements and better formulations for the resubmission of our JOSS paper.
