+++ paper/paper.md
@@ -109,7 +109,7 @@ The computer model runs using the following steps:
 
 To illustrate that the software is capable of reproducing scientific results, we first repeated an experiment from [@hintze2015punishment]. In that experiment the synergy factor in the public goods game was varied and the outcome of evolution, given that factor, determined. When the synergy factor is low, we expect defectors to win, while cooperators should thrive when synergy is high. However, in this variant of the game, agents can also punish, and punishment can modulate the response of agents to the synergy factor. The expectation was, that punishment should allow agents to cooperate at lower synergy factors, and one should observe punishment to increase around this critical point. Interestingly, agents' probabilities to cooperate and to punish evolved as expected, and punishment lowered the critical point, but the chance to punish went from 0.0 to 0.5 (drifting). \autoref{fig:Figure6} confirms both the phenomenon of drifting as well as the fact that at a synergy level of around $4$ there are hardly any defectors left. 
 
-Let us look at the shift of the critical point from 5.0 (without punishment) down to a lower level (with punishment) in a little more detail. The critical point is defined as the smallest synergy factor for which the payoff of a cooperator equals or exceeds the payoff of a defector. In [@hintze2015punishment] it was first observed that this critical point depends upon the density of punishers $\rho_{P}$ and in [@hintze2020inclusive] it was pointed out that in the absence of mutations the critical point can be expressed as  
+Let us look at the shift of the critical point from 5.0 (without punishment) down to a lower level (with punishment) in a little more detail. The critical point is defined as the smallest synergy factor for which the payoff of a cooperator equals or exceeds the payoff of a defector. In [@hintze2015punishment] it was first observed that this critical point depends upon the density of punishers $\rho_{P}$ and in [@hintze2020inclusive] it was pointed out that in the absence of mutations the critical point can be expressed as
 \begin{align} \label{criticalPointWithP}
 r_{C} = i (1 - \beta \rho_{P}) (k+1) - i + 1  - (1- i) (\beta + \gamma) \rho_{P}
 \end{align}
@@ -127,7 +127,7 @@ For example, for $\rho_{P}=0.4$ and our setting of parameters, expression (\ref{
 We also confirm a result from [@hintze2020inclusive]. Again, the synergy factor in the public goods game was varied and the outcome of evolution, given that factor, determined. In this variant of the game, there are neither fines nor costs for punishment.
 However, we impose a payoff individualism of $i=0.5$, meaning that players keep only half of their individual payoffs whereas the other half is transferred into a pool and then distributed evenly among the group members.
 Expression (\ref{criticalPointWithP}) predicts $r_{C} = 0.5 \cdot 5 - 0.5 + 1 = 3$ implying that as soon as the synergy factor exceeds 3 cooperators will benefit from a larger payoff than defectors.
-\autoref{fig:secondProof} confirms the shift of the critical point from 5.0 (classical public goods game, $i=1$) down to a synergy factor of 3 with probability to cooperate of 0.5 for that case.
+\autoref{fig:secondProof} confirms the shift of the critical point from 5.0 (classical public goods game, $i=1$) down to a synergy factor of 3 with a probability to cooperate of 0.5 for that case.
 It also displays the gradual transition from defection to cooperation in the presence of mutations reported in [@hintze2020inclusive], Figure 1.
 The probability to punish is drifting throughout the experiment as expected. 
 
@@ -135,7 +135,7 @@ The probability to punish is drifting throughout the experiment as expected.
 
 
 ## Conclusion
[…]
