### 2023-06-20 f81dcbb
Try out if the following LaTeX convention works ...
