### 2023-06-21 #4944 @jhstaudacher
RESPONSE TO THE REVIEWER:

accompanying the third resubmission of the paper

'APGG - A Modular C++ Framework for Asymmetric Public Goods Games'

by Mirko Rosenthal, David J. Richter, Falk Hübner, Jochen Staudacher and Arend Hintze

@mstimberg: Thank you very much again for your third review of our paper from April 25 and your very helpful suggestions and comments.

We worked very hard on the reproducibility issue.

>
> I am of course aware that this JOSS submission is not
> supposed to be primarily about the results in Hintze & Adami,
> but given that this replication is the only provided data
> point to show that the tool is performing its task correctly,
> I think it is necessary to be a bit pedantic about this point.
>

We completely agree with your statement. There needs to be more than one data point in order to show that our software APGG is performing its tasks correctly.

Hence we added a second of proof concept experiment and ran our first proof of concept experiment with 100,000 replicates in order to obtain a more convincing graph.

In the resubmission of our manuscript, we point out in more detail why our two experiments reproduce existing research results correctly. We apologize for the fact that our paper is now eight pages long (rather than seven pages as before).

We adjusted the wiki

https://github.com/APGG-Lab/APGG/wiki/Proof-of-Concept-Replication

accordingly. Since we ran both proof of concept experiments on a (not quite powerful) linux cluster, we are also advising our users to employ multiple cores for reproducing our graphs on the above wiki page.

We are very sorry that you were not able to reproduce our old Figure 4 (generated with only 10,000 replicates back then). It was due to the fact that the parameter 'payoffIndividualism' was set to 0.5 (rather than 1.0 for the classical public goods game). We are elaborating on these parameters in more detail in the new version of our paper, see equation (1) in our improved manuscript.

Finally, please allow me a short comment on the graph you generated and kindly included in your review from April 25. With a punishment cost gamma = 0.2, a punishment fine beta = 0.8, a group size k+1=5, your settings for the the parameter 'payoffIndividualism' i=0.5 and assuming a density of punishers (probability to punish) of 0.2 equation (1) predicts a critical point at

0.5 * (1 - 0.8 * 0.2) * 5 - 0.5 + 1  - (1-0.5) * (0.8 + 0.2) * 0.2 = 2.5

in the absence of mutations. Taking into account that we botched the offset by 0.25, this is what I am observing from your graphics. In other words: The graphics you sent us April 25 gives no indication that our software was not working correctly. Instead, we need to apologize once again that we did not make our old Figure 4 reproducible as we should have.

Furthermore, we also took care of the incorrect offsets in our plots, fixed plot.py for a missing trailing semicolon for linux and updated buildConfig.py according to your suggestions.

The two "minor issues" from your review were addressed and the paper and our wiki were adjusted accordingly.
