### 2023-06-26 #4944 @jhstaudacher
Unfortunately, our "Response to the reviewer" from June 21 contains an error. We wrote

>
> Hence we added a second of proof concept
> experiment and ran our first proof of
> concept experiment with 100,000 replicates
> in order to obtain a more convincing graph.
>

What we meant to say was the following.

Hence we added a second of proof concept experiment and ran our first proof of concept experiment for 100,000 generations with 100 replicates in order to obtain a more convincing graph.

In our paper itself, all that is correct.

We sincerely apologize for any inconvenience caused by our correction.
