### 2023-06-30 #4944 @mstimberg
Thanks to the authors for the corrections and additions to the paper and documentation. I can confirm that I can reproduce the proof of concept replications from the wiki with the given instructions, and I am now in general happy with the instructions (both in the wiki and in the paper) to obtain these results.

Regarding my earlier comment ("More generally, I am still confused about the difference in results compared to Hintze & Adami (2015)") and your reply

> In other words: The graphics you sent us April 25 gives no indication that our software was not working correctly.

Please let me state again that the research topic is not in my domain of expertise – I am therefore far from claiming that the software is calculating things incorrectly. Just to clarify my point: I was under the impression that your claim that your results replicate Hintze & Adami (2015) meant that your Figure 4 in the current manuscript was supposed to replicate Figure 6 in Hintze & Adami (2015), since both figures appeared to me to plot the same thing for the same parameters. However, your paper never refers to that specific figure, and the match between the analytical results and the simulations are convincing to me. Personally, I am still curious what leads to the differences between the two figures, but this is not relevant for this review.

**In conclusion, I am happy to recommend the software/paper for acceptance**. Below I have listed a few remaining minor points/suggestions for improvements, but from my side, there is no need for a formal back-and-forth to validate these changes, in case you decide to apply them.

### Minor issues/suggestions

#### Paper
I think the "Abstract / Summary" section title should just be "Summary"

#### Wiki
https://github.com/APGG-Lab/APGG/wiki/Build-From-Scratch
In the Linux instructions, `cd apgg` should be `cd APGG` (but it is also misleading since this is both the name of the repository and the name of the source directory, one should stay in the main folder of repository and not `cd` into the source directory)

https://github.com/APGG-Lab/APGG/wiki/ConfigValues
`MatchupGenerator` is a broken link/missing page

https://github.com/APGG-Lab/APGG/wiki/Proof-of-Concept-Replication
`Generations` is called `numGenerations` in `buildConfig.py`

https://github.com/APGG-Lab/APGG/wiki/Plotting-Data
You might consider testing/recommending newer versions? Python 3.6 reached end-of-life in December 2021

https://github.com/APGG-Lab/APGG/wiki/Proof-of-Concept-Replication
You might consider moving the remark that Proof of concept 1 is already defined in the `buildConfig.py` file to the top.

#### Software
The code in https://github.com/APGG-Lab/APGG/blob/69ab1534f475ad23730e3ade10caa68fab00756a/build/experiments/combineExperiments.py#L72-L77 seems to be an unnecessarily complex alternative to simply writing `stdDev = np.std(dfDev[groups].values)`, but there might be some subtlety I am missing
