### 2023-07-14 9d64596
JOSS paper: The "Abstract / Summary" section title should just be "Summary".
