### 2023-07-14 f43225f
JOSS paper: Update date for paper to 15 July 2023
