### 2023-07-17 0418da6
Update combineExperiments.py
Updated mean calculation with already available function from numpy.
