### 2023-07-17 9ec94e3
updated python scripts to run on newest versions of python, numpy, matplotlib and pandas, using deprecated function
