### 2023-07-17 dfb30c1
updated deprecated functions to updated ones
