### 2023-08-29 #4944 @Nikoleta-v3
Thank for your patience @jhstaudacher and my apologises for this delay. I have been on a break.

Could you please fix the following minor typos in the paper:

- Statement of need: `Studying evolution in a biological` remove `a` (Line 52)
- Acknowledgements: `improved` should be `improve` (Line 189)
- Introduction / Literature Review: ` by providing a modular framework, that is` remove comma (Line 30)
- Workflow: `The expectation was, that punishment` remove comma (Line 135)

Once you have fixed the above typos please:

1. (if necessary) Make a new tagged release of your software, and list the version tag of the archived version here. Or let me know if I should just use the release 1.1.0
2. Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
3. Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID. Please list the DOI of the archived version here.
