### 2023-08-29 c0d50a8
Fixed typos in paper.md
