### 2023-09-11 #4944 @jhstaudacher
Hey @Nikoleta-v3 Thank you very much for your feedback from August 29 and September 11 as well as for your patience. The semester has not yet restarted for me here in Bavaria ...

We took care of all the minor typos on your list. Version 1.1.2 is the version of our software APGG accepted by the reviewers. We archived APGG version 1.1.2 via Zenodo. The DOI of the archived version is 10.5281/zenodo.8334926

Thank you so much for all your efforts in editing our paper.

My coauthors and me are looking forward to the next steps.
