### 2023-09-13 a17a4f0
Create CITATION.cff
