### 2023-09-13 #4944 @jhstaudacher
@danielskatz Just to say that the PDF of our paper now downloads properly. As far as I can see, everything is fine (-- but you are the expert). Thank you very much again for all your support and efforts.
